Move test into general file
