+++ tests/index.test.tsx
@@ -1,4 +1,4 @@
-import { render, screen } from '@testing-library/react'
+import {fireEvent, render, screen} from '@testing-library/react'
 import { expectTypeOf } from 'expect-type'
 import React from 'react'
 import { describe, expect, it } from 'vitest'
@@ -249,3 +249,35 @@ describe('render <JsonViewer/> with props', () => {
     ]}/>)
   })
 })
+
+describe('Expand elements by click on dots', () => {
+  it('render', () => {
+    const {container, rerender} = render(
+        <JsonViewer
+            rootName={false}
+            value={['string1', 'string2']}
+            defaultInspectDepth={0}
+        />
+    )
+
+    let elements = container.getElementsByClassName('data-object-body');
+    expect(elements.length).eq(1)
+    expect(elements[0].textContent).eq('...')
+    fireEvent.click(elements[0])
+
+    rerender(
+        <JsonViewer
+            rootName={false}
+            value={['string1', 'string2']}
+            defaultInspectDepth={0}
+        />
+    )
+    elements = container.getElementsByClassName('data-object-body');
+    expect(elements.length).eq(0)
+
+    elements = container.getElementsByClassName('data-object');
+    expect(elements.length).eq(1)
+    expect(elements[0].children.length).eq(2)
+  })
+})
+
